Refactor tasks tests
